[*] fixed type bug

+++ couriercxx/connector/IConnectorBase.cpp
@@ -8,9 +8,9 @@
 
 #include "IConnectorBase.h"
 
-const uint8_t IConnectorBase::ERR_DEFAULT		= -1;
-const uint8_t IConnectorBase::ERR_ALREADY_OPEN	= -2;
-const uint8_t IConnectorBase::ERR_NOT_OPEN		= -3;
+const int IConnectorBase::ERR_DEFAULT		= -1;
+const int IConnectorBase::ERR_ALREADY_OPEN	= -2;
+const int IConnectorBase::ERR_NOT_OPEN		= -3;
 
 IConnectorBase::IConnectorBase() { }
