+++ couriercxx/connector/IConnectorBase.h
@@ -16,9 +16,9 @@
  */
 class IConnectorBase {
 public:
-	static const uint8_t ERR_DEFAULT;
-	static const uint8_t ERR_ALREADY_OPEN;
-	static const uint8_t ERR_NOT_OPEN;
+	static const int ERR_DEFAULT;
+	static const int ERR_ALREADY_OPEN;
+	static const int ERR_NOT_OPEN;
 
 	/**
 	 * Empty constructor
